refactor(post.type): Change filename to post.type
