+++ web/src/helpers/fetch.ts
@@ -2,7 +2,7 @@
  * Functions to Fetch data from the API
  */
 
-import { Post } from "../types/post";
+import type { Post } from "../types/post.type";
 
 export async function fetchPosts() {
   const res = await fetch("/api/Allposts");
