+++ web/src/types/post.type.ts
